data/requirements.txt
azure-cosmos>=4.5.1
GitPython>=3.1.40
python-dotenv>=1.0.0